Fixed bug: camera from exif does not contain maker.
When camera from exif did not contain maker, the models were not
compared at all. For example, Panasonic Lumix did not work, since
Panasonic does not write maker: Panasonic into the exif.

We fixed this in the way, that when camera from exif does not contain
maker, the models are compared right away. This could potentially cause
that wrong sensor will be selected in case two makers are making
models with the same name, but this is fairly unprobable.
Furthermore, when the exif camera does contain maker,
the process is the same as before.